pyproject.toml
[project]
name = "ollama-cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "crewai[tools]>=1.3.0",
    "crewai-tools>=1.3.0",
    "ddgs>=9.6.1",
    "fastapi>=0.120.4",
    "jinja2>=3.1.6",
    "langchain-community>=0.4.1",
    "ollama>=0.6.0",
    "python-dotenv>=1.2.1",
    "litellm>=1.79.0",
    "rich>=14.2.0",
    "uvicorn>=0.38.0",
    "scrapegraph-py>=1.38.0",
]
